+++ iceoryx_hoofs/platform/mac/source/socket.cpp
@@ -17,6 +17,24 @@
 #include "iceoryx_hoofs/platform/socket.hpp"
 #include <unistd.h>
 
+#include <thread>
+
+static struct timeval getTimeoutOfSocket(int sockfd, int option_name)
+{
+    struct timeval tv;
+    socklen_t optionLength = sizeof(struct timeval);
+    if (getsockopt(sockfd, SOL_SOCKET, option_name, &tv, &optionLength) == -1)
+    {
+        return {0, 0};
+    }
+    return tv;
+}
+
+static void sleepFor(struct timeval& tv)
+{
+    std::this_thread::sleep_for(std::chrono::seconds(tv.tv_sec) + std::chrono::microseconds(tv.tv_usec));
+}
+
 int iox_bind(int sockfd, const struct sockaddr* addr, socklen_t addrlen)
 {
     return bind(sockfd, addr, addrlen);
@@ -35,12 +53,26 @@ int iox_setsockopt(int sockfd, int level, int optname, const void* optval, sockl
 ssize_t
 iox_sendto(int sockfd, const void* buf, size_t len, int flags, const struct sockaddr* dest_addr, socklen_t addrlen)
 {
-    return sendto(sockfd, buf, len, flags, dest_addr, addrlen);
+    auto timeout = getTimeoutOfSocket(sockfd, SO_SNDTIMEO);
+    ssize_t sentBytes = sendto(sockfd, buf, len, flags, dest_addr, addrlen);
+    if (sentBytes <= 0 && timeout.tv_sec != 0 && timeout.tv_usec != 0)
+    {
+        sleepFor(timeout);
+        return sendto(sockfd, buf, len, flags, dest_addr, addrlen);
+    }
+    return sentBytes;
 }
 
 ssize_t iox_recvfrom(int sockfd, void* buf, size_t len, int flags, struct sockaddr* src_addr, socklen_t* addrlen)
 {
-    return recvfrom(sockfd, buf, len, flags, src_addr, addrlen);
+    auto timeout = getTimeoutOfSocket(sockfd, SO_RCVTIMEO);
+    ssize_t receivedBytes = recvfrom(sockfd, buf, len, flags, src_addr, addrlen);
+    if (receivedBytes <= 0 && timeout.tv_sec != 0 && timeout.tv_usec != 0)
+    {
+        sleepFor(timeout);
+        return recvfrom(sockfd, buf, len, flags, src_addr, addrlen);
+    }
+    return receivedBytes;
 }
 
 int iox_connect(int sockfd, const struct sockaddr* addr, socklen_t addrlen)